# docs/api/openapi-specification.yaml
openapi: 3.0.0
info:
  title: Portfolio Site REST API
  version: 2.0.0
  description: |
    Comprehensive REST API for Hugo-based portfolio site with content management,
    analytics, webhooks, and GraphQL integration.
    
    ## Features
    - Full CRUD operations for all content types
    - Advanced search and filtering
    - Real-time analytics and reporting
    - Webhook system for integrations
    - GraphQL endpoint for flexible queries
    - Multi-language content support
    - Image optimization and management
    - SEO and metadata management
    
  contact:
    name: API Support
    email: support@portfolio.com
    url: https://portfolio.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.portfolio.com/v2
    description: Production server
  - url: https://staging-api.portfolio.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /content/portfolio:
    get:
      tags: [Content - Portfolio]
      summary: List portfolio items
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Language'
        - name: category
          in: query
          schema:
            type: string
            enum: [web-development, language-learning, education, tools]
        - name: technology
          in: query
          schema:
            type: array
            items:
              type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Portfolio items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPortfolioItems'

    post:
      tags: [Content - Portfolio]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioRequest'
      responses:
        '201':
          description: Portfolio item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'
  /content/portfolio/{id}:
    get:
      tags: [Content - Portfolio]
      summary: Get portfolio item by ID
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Portfolio item retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'

    put:
      tags: [Content - Portfolio]
      summary: Update portfolio item
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePortfolioRequest'
      responses:
        '200':
          description: Portfolio item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'

    delete:
      tags: [Content - Portfolio]
      summary: Delete portfolio item
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '204':
          description: Portfolio item deleted successfully

  # Search and Filtering
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ContentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Content identifier
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Language:
      name: lang
      in: query
      schema:
        type: string
        enum: [en, es]
        default: en
      description: Content language
  schemas:
...
    PortfolioItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [web-development, language-learning, education, tools]
        technologies:
          type: array
          items:
            type: string
        featured:
          type: boolean
        images:
          type: array
          items:
            type: string
            format: uri
        liveUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, completed, archived]
        language:
          type: string
          enum: [en, es]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seo:
          $ref: '#/components/schemas/SEOMetadata'
    CreatePortfolioRequest:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
        category:
          type: string
          enum: [web-development, language-learning, education, tools]
        technologies:
          type: array
          items:
            type: string
        featured:
          type: boolean
          default: false
        liveUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, completed, archived]
          default: active
        language:
          type: string
          enum: [en, es]
          default: en
        seo:
          $ref: '#/components/schemas/SEOMetadataInput'
    UpdatePortfolioRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
          maxLength: 500
        content:
          type: string
        category:
          type: string
          enum: [web-development, language-learning, education, tools]
        technologies:
          type: array
          items:
            type: string
        featured:
          type: boolean
        liveUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, completed, archived]
        seo:
          $ref: '#/components/schemas/SEOMetadataInput'
...
    PaginatedPortfolioItems:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    SEOMetadata:
      type: object
      properties:
        metaTitle:
          type: string
          maxLength: 60
        metaDescription:
          type: string
          maxLength: 160
        metaKeywords:
          type: array
          items:
            type: string
        openGraph:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
            type:
              type: string
              default: article
        twitter:
          type: object
          properties:
            card:
              type: string
              default: summary_large_image
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
        structuredData:
          type: object
    SEOMetadataInput:
      type: object
      properties:
        metaTitle:
          type: string
          maxLength: 60
        metaDescription:
          type: string
          maxLength: 160
        metaKeywords:
          type: array
          items:
            type: string
        openGraph:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
        twitter:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
...